requirements.txt
appdirs==1.4.3
Django==1.11
django-cors-headers==1.1.0
django-cors-middleware==1.3.1
djangorestframework==3.4.0
gunicorn==19.7.1
mysqlclient==1.3.10
packaging==16.8
pyparsing==2.2.0
pytz==2017.2
six==1.10.0
